Local login system.

+++ HotelReservations/HotelReservations.Shared/Model/Room.cs
@@ -3,6 +3,6 @@
     public class Room
     {
         public long Id { get; set; }
-        public long MaxOccupancy { get; set; }
+        public int MaxOccupancy { get; set; }
     }
 }
